update screen file structure
