+++ src/screens/onboarding_1/components/OnboardingPanel.tsx
@@ -1,10 +1,10 @@
-import { StackNavigationProp } from "@react-navigation/stack";
-import CTAButton from "@src/components/CTAButton";
-import { RootStackParamList } from "@src/navigation/RootNavigator";
 import { useWindowDimensions } from "react-native";
 import { View, Text } from "react-native";
+import { StackNavigationProp } from "@react-navigation/stack";
+import { RootStackParamList } from "@src/navigation/RootNavigator";
 import { DeviceType, getDeviceType } from "@src/utils/deviceUtils";
 import { styles } from "@src/screens/onboarding_1/styles";
+import CTAButton from "@src/components/CTAButton";
 
 type NavigationProp = StackNavigationProp<RootStackParamList, "Onboarding_1">;
 export default function OnboardingPanel({
